% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{contracts}
\alias{contracts}
\title{Contracts for projects financed by the World Bank}
\format{A data frame with 4340 observations of 7 variables}
\usage{
data(contracts)
}
\description{
A dataset about contracts, with the following variables about each contract
}
\details{
\itemize{
\item foo. bar baz
}
}
\keyword{datasets}

