% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lowest-bidder-not-selected.R
\name{lowest.bidder}
\alias{lowest.bidder}
\title{Run the lowest bidder non-selection detector}
\usage{
lowest.bidder(contract)
}
\arguments{
\item{contract}{A data frame of bids on one contract}
}
\value{
A vector with the contract and statistics about bid selections
}
\description{
Run the lowest bidder non-selection detector
}

