% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.contract}
\alias{plot.contract}
\title{Plot the prices of bids on a contract to check for suspicious bid-rejections.}
\usage{
\method{plot}{contract}(contract)
}
\arguments{
\item{contract}{A data frame containing the bids for a single contract.}
}
\value{
A ggplot2 plot, or NULL (if the contract is missing data)
}
\description{
Make a plot in the style of Francis's conceptual drawing of the
different bid prices for the purpose of finding suspicious
rejections of low bidders.
}

