% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.lowest.bidder}
\alias{plot.lowest.bidder}
\title{Draw a plot about bid selections with a focus on detection low bidder non-selection}
\usage{
\method{plot}{lowest.bidder}(contracts)
}
\arguments{
\item{contracts}{A data frame of contracts}
}
\value{
A ggplot2 plot
}
\description{
Draw a plot about bid selections with a focus on detection low bidder non-selection
}

