% Generated by roxygen2 (4.0.1): do not edit by hand
\name{round.numbers}
\alias{round.numbers}
\title{Count round numbers in bid prices.}
\usage{
\method{round}{numbers}(bids)
}
\arguments{
\item{bids}{Data frame of bids, with the columns "currency" and "contract.number"}
}
\value{
Data frame of contracts, with the columns "round.bids", "total.bids", and "main.currency".
}
\description{
Unusually round bid prices raise suspicioun of bid rigging.
Given a data frame of bids, this function counts how many of
those bids are round. It also counts the total number of
bids and checks which currency is most common for the bids.
The currency information is mainly helpful for debugging.
}

