% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/contract-splitting.R
\name{standardize.prices}
\alias{standardize.prices}
\title{Within a particular data frame, produce a price.standardized
column as the standardized price.amount column. Ignore NAs.}
\usage{
standardize.prices(df)
}
\description{
Within a particular data frame, produce a price.standardized
column as the standardized price.amount column. Ignore NAs.
}

