% Generated by roxygen2 (4.0.1): do not edit by hand
\name{very.round}
\alias{very.round}
\title{Find suspicious roundness patterns in contracts.}
\usage{
very.round(bids)
}
\arguments{
\item{bids}{Data frame of bids, with the columns "currency" and "contract.number"}
}
\value{
Data frame of contracts, with the columns "round.bids", "total.bids", and "main.currency"
}
\description{
Unusually round bid prices raise suspicioun of bid rigging.
Given a data frame of bids, this function counts how many of
those bids are round. It also counts the total number of
bids and checks which currency is most common for the bids.
(That is, it calls `round.numbers`.) Then it selects contracts
with unusually high degrees of roundness.
}

