% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{bids}
\alias{bids}
\title{Bids on contracts for projects financed by the World Bank}
\format{A data frame with 14680 observations of 6 variables}
\usage{
data(bids)
}
\description{
A dataset about bids on contracts, with multiple bidders per contract
and with the following variables about each bid
}
\details{
\itemize{
\item foo. bar baz
}
}
\keyword{datasets}

