% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/price-roundness.R
\name{is.round}
\alias{is.round}
\title{Determine whether a bid price (or any number) is "round".}
\usage{
is.round(n)
}
\arguments{
\item{number}{The number or numbers to test for roundness}
}
\value{
Whether the number is round
}
\description{
This function tries to determine whether a bid price is
round. A bid price is considered round if it has more trailing
zeroes than other digits.
}

