% Generated by roxygen2 (4.0.1): do not edit by hand
\name{standardize.methods}
\alias{standardize.methods}
\title{Look for contract splitting, that is, repeated awards
below procurement thresholds, possibly to the same contractor.
Standardize the names of the procurement and selection methods}
\usage{
standardize.methods(contracts)
}
\arguments{
\item{contracts}{a data frame with method.procurement and method.selection columns}
}
\value{
the data frame with standardized method.procurement and method.selection columns
}
\description{
Look for contract splitting, that is, repeated awards
below procurement thresholds, possibly to the same contractor.
Standardize the names of the procurement and selection methods
}

