% Generated by roxygen2 (4.0.1): do not edit by hand
\name{standardize.prices}
\alias{standardize.prices}
\title{Within a particular data frame, produce a price.standardized
column as the standardized price.amount column. Ignore NAs.}
\usage{
standardize.prices(df)
}
\description{
Within a particular data frame, produce a price.standardized
column as the standardized price.amount column. Ignore NAs.
}

