% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/contract-splitting.R
\name{strange.prices}
\alias{strange.prices}
\title{Detect strange prices}
\usage{
strange.prices(contracts)
}
\arguments{
\item{contracts}{a data frame with contract-level data}
}
\value{
a data frame of project-level data with the newly computed statistics
}
\description{
Use Pearson kurtosis and Kolmogorov-Smirnov statistics to
look for unusual patterns in contract prices, by project.
}

